requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
azure-storage-blob
msal
pyodbc